+++ Blackjack/Model/Royals.swift
@@ -1,0 +1,16 @@
+//
+//  Royals.swift
+//  Blackjack
+//
+//  Created by Jack Munkonge on 20/11/2019.
+//  Copyright © 2019 MobileDev. All rights reserved.
+//
+
+import Foundation
+
+struct Royals {
+    static let jack = "jack"
+    static let queen = "queen"
+    static let king = "king"
+    static let ace = "ace"
+}
