+++ Blackjack/Model/Suits.swift
@@ -1,0 +1,16 @@
+//
+//  Suits.swift
+//  Blackjack
+//
+//  Created by Jack Munkonge on 20/11/2019.
+//  Copyright © 2019 MobileDev. All rights reserved.
+//
+
+import Foundation
+
+struct Suits {
+    static let spades = "spades"
+    static let clubs = "clubs"
+    static let hearts = "hearts"
+    static let diamonds = "diamonds"
+}
